check-runner/pyproject.toml
[project]
name = "check-runner"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "asyncpg==0.30.0",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
    "mygeotab>=0.9.4",
    "psycopg2-binary>=2.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -ra"
asyncio_mode = "auto"
